requirements.txt
matplotlib==3.3.1
numpy==1.19.3
pycaret-nightly[full]
pandas==1.2.0
scikit-learn==0.23.2
seaborn==0.11.1
streamlit==0.75.0
streamlit-pandas-profiling==0.1.1
Pillow==8.0.1
pandas-profiling==2.10.0
hpbandster==0.7.4
ConfigSpace==0.4.17
xlrd==2.0.1
xlsxwriter==1.3.7
openpyxl==3.0.6